Changes: Added services folder for any scraping service files and wrote a script for ABV.
Required future steps:
Clear popular_sites.txt from sites that do not have a register/forgotten password page or a method to check if an account is registered or not.

Add a big file for printing out results and start writing scrapers around it. Discuss how this can be approached best. (GUI/Terminal?)